requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
python-jose[cryptography]
bcrypt
passlib
python-multipart
langchain
langchain-ollama
langchain-chroma
langchain-community
langchain-text-splitters
chromadb
pypdf
python-docx
sse-starlette
torch
rank-bm25
sentence-transformers
scikit-learn
bleach
slowapi
watchdog
schedule